+++ poke-infirmerie/src/main/java/fr/pantheonsorbonne/ufr27/miage/model/Dresseur.java
@@ -1,5 +1,6 @@
 package fr.pantheonsorbonne.ufr27.miage.model;
 
+
 import jakarta.persistence.*;
 
 import java.util.List;
