+++ poke-infirmerie/src/main/java/fr/pantheonsorbonne/ufr27/miage/model/TreatmentSession.java
@@ -1,5 +1,6 @@
 package fr.pantheonsorbonne.ufr27.miage.model;
 
+
 import jakarta.persistence.*;
 
 import java.sql.Time;
