+++ poke-school/src/main/java/fr/pantheonsorbonne/ufr27/miage/model/SchoolTicket.java
@@ -1,0 +1,37 @@
+package fr.pantheonsorbonne.ufr27.miage.model;
+
+import jakarta.persistence.*;
+
+@Entity
+public class SchoolTicket {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "idSchoolTicket", nullable = false)
+    private Integer idSchoolTicket;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    private Dresseur dresseur;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    private Pokemon pokemon;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    private SchoolSession schoolSession;
+
+    public Dresseur getDresseur() {
+        return dresseur;
+    }
+
+    public void setDresseur(Dresseur dresseur) {
+        this.dresseur = dresseur;
+    }
+
+    public Pokemon getPokemon() {
+        return pokemon;
+    }
+
+    public void setPokemon(Pokemon pokemon) {
+        this.pokemon = pokemon;
+    }
+}
