+++ poke-school/src/main/resources/import.sql
@@ -1,0 +1,4 @@
+REPLACE INTO `SchoolSession` (`idSchoolSession`, `pokescoreGain`, `priceSchoolSession`, `timeSchoolSession`) VALUES
+                                                                                                                (1, 20, 40, '00:00:06.000000'),
+                                                                                                                (2, 40, 80, '00:00:06.000000'),
+                                                                                                                (3, 60, 120, '00:00:06.000000');
